refactor: changed the receiveTokensStore from js to ts
