+++ .gitignore
@@ -31,3 +31,10 @@ yarn-error.log*
 *.ntvs*
 *.njsproj
 *.sln
+
+# Package files
+package-lock.json
+package.json
+quasar.extensions.json
+jsconfig.json
+vitest.config.js
