+++ src/stores/index.ts
@@ -1,5 +1,5 @@
 import { store } from "quasar/wrappers";
-import { createPinia } from "pinia";
+import { createPinia, Pinia } from "pinia";
 
 /*
  * If not building with SSR mode, you can
@@ -10,7 +10,7 @@ import { createPinia } from "pinia";
  * with the Store instance.
  */
 
-export default store((/* { ssrContext } */) => {
+export default store((/* { ssrContext }: { ssrContext?: any } */): Pinia => {
   const pinia = createPinia();
 
   // You can add Pinia plugins here
